pyproject.toml
[tool.poetry]
name = "ipfabric"
version = "v5.0.9"
description = "Python package for interacting with IP Fabric"
authors = [
    "Justin Jeffery <justin.jeffery@ipfabric.io>",
    "Cristian Cordero <cristian.cordero@ipfabric.io>",
    "Community Fabric <communityfabric@ipfabric.io>"
]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/community-fabric/python-ipfabric"
repository = "https://github.com/community-fabric/python-ipfabric"
keywords = ["ipfabric", "ip-fabric", "community-fabric"]

[tool.poetry.urls]
"IP Fabric" = "https://ipfabric.io/"
"Changelog" = "https://github.com/community-fabric/python-ipfabric/blob/main/CHANGELOG.md"

[tool.poetry.dependencies]
python = "^3.7.1"
httpx = "^0.23.0"
python-dateutil = "^2.8.2"
pydantic = "^1.8.2"
pytz = "^2022.1"
python-dotenv = "^0.21"
pandas = {version = "1.3.0", optional = true}
openpyxl = {version = "^3.0.9", optional = true}
tabulate = {version = "^0.8.9",  optional = true}
ipfabric-httpx-auth = "^5.0.0"
python-json-logger = {version = "^2.0.4",  optional = true}

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
flake8 = "^4.0.1"
black = "^22.3"

[tool.poetry.extras]
examples = ["tabulate", "pandas", "openpyxl", "python-json-logger"]

[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | examples
  )/
  | settings.py     # This is where you define files that should not be stylized by black
                     # the root of the project
)
'''